+++ wab/create/Create_ListView.cpp
@@ -113,7 +113,7 @@ void ListView::Create( HWND parent )
 		GetActualWidth(),
 		GetActualHeight(), 
 		parent,
-		(HMENU)GetID(),
+		(HMENU)(uint64_t)GetID(),
 		0,
 		0
 	);
