+++ wab/create/Create_Richtext.cpp
@@ -155,7 +155,7 @@ void Richtext::Create( HWND parent )
 		GetActualWidth(),
 		GetActualHeight(), 
 		parent,
-		(HMENU)GetID(),
+		(HMENU)(uint64_t)GetID(),
 		0,
 		0
 	);
