+++ wab/create/Create_Treeview.cpp
@@ -73,7 +73,7 @@ void Treeview::Create( HWND parent )
 		GetActualWidth(),
 		GetActualHeight(),
 		parent,
-		(HMENU)GetID(),
+		(HMENU)(uint64_t)GetID(),
 		0,
 		0
 	);
